roo-code/backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
requests==2.26.0
websocket-client==1.2.1